refactor(api) migrate http-errors and pre-response-utils
