+++ api/config/server-setup-error-handling.js
@@ -1,4 +1,3 @@
-import * as preResponseUtils from '../lib/application/pre-response-utils.js';
 import { certificationDomainErrorMappingConfiguration } from '../src/certification/shared/application/http-error-mapper-configuration.js';
 import { devcompDomainErrorMappingConfiguration } from '../src/devcomp/application/http-error-mapper-configuration.js';
 import { evaluationDomainErrorMappingConfiguration } from '../src/evaluation/application/http-error-mapper-configuration.js';
@@ -7,7 +6,7 @@ import { organizationalEntitiesDomainErrorMappingConfiguration } from '../src/or
 import { prescriptionDomainErrorMappingConfiguration } from '../src/prescription/shared/application/http-error-mapper-configuration.js';
 import { schoolDomainErrorMappingConfiguration } from '../src/school/application/http-error-mapper-configuration.js';
 import { domainErrorMapper } from '../src/shared/application/domain-error-mapper.js';
-import * as sharedPreResponseUtils from '../src/shared/application/pre-response-utils.js';
+import * as preResponseUtils from '../src/shared/application/pre-response-utils.js';
 import { teamDomainErrorMappingConfiguration } from '../src/team/application/http-error-mapper-configuration.js';
 
 const setupErrorHandling = function (server) {
@@ -25,7 +24,6 @@ const setupErrorHandling = function (server) {
   domainErrorMapper.configure(configuration);
 
   server.ext('onPreResponse', preResponseUtils.handleDomainAndHttpErrors);
-  server.ext('onPreResponse', sharedPreResponseUtils.handleDomainAndHttpErrors);
 };
 
 export { setupErrorHandling };
