+++ api/lib/application/certification-center-memberships/certification-center-membership-controller.js
@@ -1,8 +1,8 @@
+import { BadRequestError, ForbiddenError } from '../../../src/shared/application/http-errors.js';
 import * as requestResponseUtils from '../../../src/shared/infrastructure/utils/request-response-utils.js';
 import { usecases } from '../../domain/usecases/index.js';
 import { getCertificationCenterId } from '../../infrastructure/repositories/certification-center-membership-repository.js';
 import * as certificationCenterMembershipSerializer from '../../infrastructure/serializers/jsonapi/certification-center-membership-serializer.js';
-import { BadRequestError, ForbiddenError } from '../http-errors.js';
 
 const disableFromPixAdmin = async function (request, h, dependencies = { requestResponseUtils }) {
   const certificationCenterMembershipId = request.params.id;
