+++ api/lib/application/target-profiles/index.js
@@ -1,8 +1,8 @@
 import Joi from 'joi';
 
+import { BadRequestError, sendJsonApiError } from '../../../src/shared/application/http-errors.js';
 import { securityPreHandlers } from '../../../src/shared/application/security-pre-handlers.js';
 import { identifiersType } from '../../../src/shared/domain/types/identifiers-type.js';
-import { BadRequestError, sendJsonApiError } from '../http-errors.js';
 import { targetProfileController } from './target-profile-controller.js';
 
 const register = async function (server) {
