+++ api/lib/domain/usecases/disable-certification-center-membership-from-pix-certif.js
@@ -1,5 +1,5 @@
+import { ForbiddenError } from '../../../src/shared/application/http-errors.js';
 import { NotFoundError } from '../../../src/shared/domain/errors.js';
-import { ForbiddenError } from '../../application/http-errors.js';
 
 const disableCertificationCenterMembershipFromPixCertif = async function ({
   certificationCenterMembershipId,
