+++ api/src/certification/flash-certification/application/scenario-simulator-controller.js
@@ -2,9 +2,9 @@ import { Readable } from 'node:stream';
 
 import _ from 'lodash';
 
-import { HttpErrors } from '../../../../lib/application/http-errors.js';
 import { parseCsv } from '../../../../scripts/helpers/csvHelpers.js';
 import { pickChallengeService } from '../../../evaluation/domain/services/pick-challenge-service.js';
+import { HttpErrors } from '../../../shared/application/http-errors.js';
 import { random } from '../../../shared/infrastructure/utils/random.js';
 import { extractLocaleFromRequest } from '../../../shared/infrastructure/utils/request-response-utils.js';
 import { pickAnswerStatusService } from '../../shared/domain/services/pick-answer-status-service.js';
