+++ api/src/shared/application/http-errors.js
@@ -184,6 +184,7 @@ export {
   PreconditionFailedError,
   sendJsonApiError,
   ServiceUnavailableError,
+  SessionPublicationBatchError,
   TooManyRequestsError,
   UnauthorizedError,
   UnprocessableEntityError,
