+++ api/tests/evaluation/integration/infrastructure/repositories/badge-criteria-repository_test.js
@@ -1,6 +1,6 @@
-import { BadRequestError } from '../../../../../lib/application/http-errors.js';
 import BadgeCriterion from '../../../../../src/evaluation/domain/models/BadgeCriterion.js';
 import * as badgeCriteriaRepository from '../../../../../src/evaluation/infrastructure/repositories/badge-criteria-repository.js';
+import { BadRequestError } from '../../../../../src/shared/application/http-errors.js';
 import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
 import { SCOPES } from '../../../../../src/shared/domain/models/BadgeDetails.js';
 import { catchErr, databaseBuilder, expect, knex } from '../../../../test-helper.js';
