+++ api/tests/integration/application/pre-response-utils_test.js
@@ -8,8 +8,8 @@ import {
   PreconditionFailedError,
   UnauthorizedError,
   UnprocessableEntityError,
-} from '../../../lib/application/http-errors.js';
-import { handleDomainAndHttpErrors } from '../../../lib/application/pre-response-utils.js';
+} from '../../../src/shared/application/http-errors.js';
+import { handleDomainAndHttpErrors } from '../../../src/shared/application/pre-response-utils.js';
 import { expect, hFake } from '../../test-helper.js';
 
 describe('Integration | Application | PreResponse-utils', function () {
