+++ api/tests/shared/unit/application/pre-response-utils_test.js
@@ -1,7 +1,6 @@
-import { BaseHttpError } from '../../../lib/application/http-errors.js';
-import { handleDomainAndHttpErrors } from '../../../lib/application/pre-response-utils.js';
-import { DomainError } from '../../../src/shared/domain/errors.js';
-import { expect, hFake, sinon } from '../../test-helper.js';
+import { BaseHttpError } from '../../../../src/shared/application/http-errors.js';
+import { handleDomainAndHttpErrors } from '../../../../src/shared/application/pre-response-utils.js';
+import { expect, hFake, sinon } from '../../../test-helper.js';
 
 describe('Unit | Application | PreResponse-utils', function () {
   describe('#handleDomainAndHttpErrors', function () {
@@ -28,18 +27,6 @@ describe('Unit | Application | PreResponse-utils', function () {
       // then
       expect(typeof response).to.be.equal('symbol');
       expect(response.toString()).to.be.equal(expectedString);
-    });
-
-    it('should not manage (src/shared) DomainError', async function () {
-      const request = {
-        response: new DomainError('Error message'),
-      };
-
-      // when
-      await handleDomainAndHttpErrors(request, hFake, { errorManager });
-
-      // then
-      expect(errorManager.handle).not.to.have.been.calledWithExactly(request, hFake, request.response);
     });
 
     it('should manage BaseHttpError', async function () {
