+++ api/tests/shared/unit/validate_test.js
@@ -2,7 +2,7 @@ import lodash from 'lodash';
 
 const { pick } = lodash;
 
-import { BadRequestError } from '../../../lib/application/http-errors.js';
+import { BadRequestError } from '../../../src/shared/application/http-errors.js';
 import { handleFailAction } from '../../../src/shared/validate.js';
 import { expect, hFake } from '../../test-helper.js';
